Prototype de menu termine
Prototype terminé ayant la possibilité de changer de frame, vers celle du jeu, ou de quitter.

A integrer plus tard : Difficulte et Highscore.
